refactor: mock data를 /mocks/data로 이동 (#141)
