+++ src/mocks/data/mockFilterCategories.ts
