+++ src/mocks/index.ts
@@ -1,5 +1,14 @@
-export { default as MOCK_FILTER_CATEGORIES } from './mockFilterCategories'
-export { default as MOCK_GAMES, type MockGame } from './mockGames'
-export { DUMMY_GAMES } from './dummyGames'
-export { MOCK_TAGS } from './mockTags'
-export { MOCK_GENRES } from './mockGenres'
+import { DUMMY_GAMES } from './data/dummyGames'
+import { default as MOCK_FILTER_CATEGORIES } from './data/mockFilterCategories'
+import { default as MOCK_GAMES, type MockGame } from './data/mockGames'
+import { MOCK_GENRES } from './data/mockGenres'
+import { MOCK_TAGS } from './data/mockTags'
+
+export {
+  DUMMY_GAMES,
+  MOCK_FILTER_CATEGORIES,
+  MOCK_GAMES,
+  MOCK_GENRES,
+  MOCK_TAGS,
+  MockGame,
+}
